Conflicts:
	scilab/Makefile.am
	scilab/Makefile.in
	scilab/configure
	scilab/etc/modules.xml.in
	scilab/etc/modules.xml.vc
	scilab/modules/core/Makefile.am
	scilab/modules/core/Makefile.in

+++ scilab/modules/elementary_functions/sci_gateway/cpp/sci_imult.cpp
@@ -77,18 +77,18 @@ types::Function::ReturnValue sci_imult(types::typed_list &in, int _iRetCount, ty
         {
             for (int i = 0 ; i < nonZeros ; i++)
             {
-                std::complex<double> cplxIn = pSparseIn->getImg(pRows[i] - 1, pCols[i] - 1);
+                std::complex<double> cplxIn = pSparseIn->getImg((int)pRows[i] - 1, (int)pCols[i] - 1);
                 std::complex<double> cplxOut(cplxIn.imag() * -1, cplxIn.real());
-                pSparseOut->set(pRows[i] - 1, pCols[i] - 1, cplxOut);
+                pSparseOut->set((int)pRows[i] - 1, (int)pCols[i] - 1, cplxOut);
             }
         }
         else
         {
             for (int i = 0 ; i < nonZeros ; i++)
             {
-                double dReal = pSparseIn->get(pRows[i] - 1, pCols[i] - 1);
+                double dReal = pSparseIn->get((int)pRows[i] - 1, (int)pCols[i] - 1);
                 std::complex<double> cplxOut(0, dReal);
-                pSparseOut->set(pRows[i] - 1, pCols[i] - 1, cplxOut);
+                pSparseOut->set((int)pRows[i] - 1, (int)pCols[i] - 1, cplxOut);
             }
         }
